+++ CHANGELOG.md
@@ -1,6 +1,6 @@
 # Changelog
 
-All notable changes to `laravel-repository` will be documented in this file
+All notable changes to `laravel-repository` will be documented in this file.
 
 ## next