+++ CHANGELOG.md
@@ -4,6 +4,8 @@ All notable changes to `laravel-repository` will be documented in this file.
 
 ## next
 
+## 0.8.0
+
 -   feat: 1st argument to `setContext()` can be an array to get ArrayResourceContext
 -   feat: add 2nd argument to `setContext()` so automatically setAllowedWith
     equal to whatever is set in the context.